feat(#366): 관심 클래스 로딩 구현
